+++ ChangeLog.md
@@ -1,4 +1,10 @@
 # Changelog for org-to-things
 
+## v1.1.0.0
+- Added support for INACTIVE todos which get mapped to "Someday" in Things.
+
+## v1.0.0.1
+- Fixed bug where a Heading with no Todos failed to parse
+
 ## v1.0.0.0
 - Generates links to add Todos and Projects.